Missing readme instructions and fixes
Needed to `rustup target add` for wasm when getting setup. Was missing
`cwebp` on ubuntu, so added that to dependencies.

The readme says to run `cargo run` from the project root directory, but
that breaks the relative file paths in the server. Should instead change
the working dir to the server and run from there.